threw this into a package called: gophernet
