+++ gophernet/gophernet.go
@@ -1,0 +1,229 @@
+package gophernet
+
+import (
+	"errors"
+	"math"
+	"math/rand"
+	"time"
+
+	"gonum.org/v1/gonum/floats"
+	"gonum.org/v1/gonum/mat"
+)
+
+// GopherNet contains all of the information
+// that defines a trained neural network.
+type GopherNet struct {
+	config  GopherNetConfig
+	wHidden *mat.Dense
+	bHidden *mat.Dense
+	wOut    *mat.Dense
+	bOut    *mat.Dense
+}
+
+// GopherNetConfig defines our neural network
+// architecture and learning parameters.
+type GopherNetConfig struct {
+	InputNeurons  int
+	OutputNeurons int
+	HiddenNeurons int
+	NumEpochs     int
+	LearningRate  float64
+}
+
+// NewNetwork initializes a new neural network.
+func NewNetwork(config GopherNetConfig) *GopherNet {
+	return &GopherNet{config: config}
+}
+
+// train trains a neural network using backpropagation.
+func (gn *GopherNet) Train(x, y *mat.Dense) error {
+
+	// Initialize biases/weights.
+	randSource := rand.NewSource(time.Now().UnixNano())
+	randGen := rand.New(randSource)
+
+	wHidden := mat.NewDense(gn.config.InputNeurons, gn.config.HiddenNeurons, nil)
+	bHidden := mat.NewDense(1, gn.config.HiddenNeurons, nil)
+	wOut := mat.NewDense(gn.config.HiddenNeurons, gn.config.OutputNeurons, nil)
+	bOut := mat.NewDense(1, gn.config.OutputNeurons, nil)
+
+	wHiddenRaw := wHidden.RawMatrix().Data
+	bHiddenRaw := bHidden.RawMatrix().Data
+	wOutRaw := wOut.RawMatrix().Data
+	bOutRaw := bOut.RawMatrix().Data
+
+	for _, param := range [][]float64{
+		wHiddenRaw,
+		bHiddenRaw,
+		wOutRaw,
+		bOutRaw,
+	} {
+		for i := range param {
+			param[i] = randGen.Float64()
+		}
+	}
+
+	// Define the output of the neural network.
+	output := new(mat.Dense)
+
+	// Use backpropagation to adjust the weights and biases.
+	if err := gn.backpropagate(x, y, wHidden, bHidden, wOut, bOut, output); err != nil {
+		return err
+	}
+
+	// Define our trained neural network.
+	gn.wHidden = wHidden
+	gn.bHidden = bHidden
+	gn.wOut = wOut
+	gn.bOut = bOut
+
+	return nil
+}
+
+// Backpropagate completes the backpropagation method.
+func (gn *GopherNet) backpropagate(x, y, wHidden, bHidden, wOut, bOut, output *mat.Dense) error {
+
+	// Loop over the number of epochs utilizing
+	// backpropagation to train our model.
+	for i := 0; i < gn.config.NumEpochs; i++ {
+
+		// Complete the feed forward process.
+		hiddenLayerInput := new(mat.Dense)
+		hiddenLayerInput.Mul(x, wHidden)
+		addBHidden := func(_, col int, v float64) float64 { return v + bHidden.At(0, col) }
+		hiddenLayerInput.Apply(addBHidden, hiddenLayerInput)
+
+		hiddenLayerActivations := new(mat.Dense)
+		applySigmoid := func(_, _ int, v float64) float64 { return sigmoid(v) }
+		hiddenLayerActivations.Apply(applySigmoid, hiddenLayerInput)
+
+		outputLayerInput := new(mat.Dense)
+		outputLayerInput.Mul(hiddenLayerActivations, wOut)
+		addBOut := func(_, col int, v float64) float64 { return v + bOut.At(0, col) }
+		outputLayerInput.Apply(addBOut, outputLayerInput)
+		output.Apply(applySigmoid, outputLayerInput)
+
+		// Complete the backpropagation.
+		networkError := new(mat.Dense)
+		networkError.Sub(y, output)
+
+		slopeOutputLayer := new(mat.Dense)
+		applySigmoidPrime := func(_, _ int, v float64) float64 { return sigmoidPrime(v) }
+		slopeOutputLayer.Apply(applySigmoidPrime, output)
+		slopeHiddenLayer := new(mat.Dense)
+		slopeHiddenLayer.Apply(applySigmoidPrime, hiddenLayerActivations)
+
+		dOutput := new(mat.Dense)
+		dOutput.MulElem(networkError, slopeOutputLayer)
+		errorAtHiddenLayer := new(mat.Dense)
+		errorAtHiddenLayer.Mul(dOutput, wOut.T())
+
+		dHiddenLayer := new(mat.Dense)
+		dHiddenLayer.MulElem(errorAtHiddenLayer, slopeHiddenLayer)
+
+		// Adjust the parameters.
+		wOutAdj := new(mat.Dense)
+		wOutAdj.Mul(hiddenLayerActivations.T(), dOutput)
+		wOutAdj.Scale(gn.config.LearningRate, wOutAdj)
+		wOut.Add(wOut, wOutAdj)
+
+		bOutAdj, err := sumAlongAxis(0, dOutput)
+		if err != nil {
+			return err
+		}
+		bOutAdj.Scale(gn.config.LearningRate, bOutAdj)
+		bOut.Add(bOut, bOutAdj)
+
+		wHiddenAdj := new(mat.Dense)
+		wHiddenAdj.Mul(x.T(), dHiddenLayer)
+		wHiddenAdj.Scale(gn.config.LearningRate, wHiddenAdj)
+		wHidden.Add(wHidden, wHiddenAdj)
+
+		bHiddenAdj, err := sumAlongAxis(0, dHiddenLayer)
+		if err != nil {
+			return err
+		}
+		bHiddenAdj.Scale(gn.config.LearningRate, bHiddenAdj)
+		bHidden.Add(bHidden, bHiddenAdj)
+	}
+
+	return nil
+}
+
+// predict makes a prediction based on a trained
+// neural network.
+func (gn *GopherNet) Predict(x *mat.Dense) (*mat.Dense, error) {
+
+	// Check to make sure that our GopherNet value
+	// represents a trained model.
+	if gn.wHidden == nil || gn.wOut == nil {
+		return nil, errors.New("the supplied weights are empty")
+	}
+	if gn.bHidden == nil || gn.bOut == nil {
+		return nil, errors.New("the supplied biases are empty")
+	}
+
+	// Define the output of the neural network.
+	output := new(mat.Dense)
+
+	// Complete the feed forward process.
+	hiddenLayerInput := new(mat.Dense)
+	hiddenLayerInput.Mul(x, gn.wHidden)
+	addBHidden := func(_, col int, v float64) float64 { return v + gn.bHidden.At(0, col) }
+	hiddenLayerInput.Apply(addBHidden, hiddenLayerInput)
+
+	hiddenLayerActivations := new(mat.Dense)
+	applySigmoid := func(_, _ int, v float64) float64 { return sigmoid(v) }
+	hiddenLayerActivations.Apply(applySigmoid, hiddenLayerInput)
+
+	outputLayerInput := new(mat.Dense)
+	outputLayerInput.Mul(hiddenLayerActivations, gn.wOut)
+	addBOut := func(_, col int, v float64) float64 { return v + gn.bOut.At(0, col) }
+	outputLayerInput.Apply(addBOut, outputLayerInput)
+	output.Apply(applySigmoid, outputLayerInput)
+
+	return output, nil
+}
+
+// sigmoid implements the sigmoid function
+// for use in activation functions.
+func sigmoid(x float64) float64 {
+	return 1.0 / (1.0 + math.Exp(-x))
+}
+
+// sigmoidPrime implements the derivative
+// of the sigmoid function for backpropagation.
+func sigmoidPrime(x float64) float64 {
+	return x * (1.0 - x)
+}
+
+// sumAlongAxis sums a matrix along a
+// particular dimension, preserving the
+// other dimension.
+func sumAlongAxis(axis int, m *mat.Dense) (*mat.Dense, error) {
+
+	numRows, numCols := m.Dims()
+
+	var output *mat.Dense
+
+	switch axis {
+	case 0:
+		data := make([]float64, numCols)
+		for i := 0; i < numCols; i++ {
+			col := mat.Col(nil, i, m)
+			data[i] = floats.Sum(col)
+		}
+		output = mat.NewDense(1, numCols, data)
+	case 1:
+		data := make([]float64, numRows)
+		for i := 0; i < numRows; i++ {
+			row := mat.Row(nil, i, m)
+			data[i] = floats.Sum(row)
+		}
+		output = mat.NewDense(numRows, 1, data)
+	default:
+		return nil, errors.New("invalid axis, must be 0 or 1")
+	}
+
+	return output, nil
+}
