+++ main.go
@@ -2,38 +2,15 @@ package main
 
 import (
 	"encoding/csv"
-	"errors"
 	"fmt"
 	"log"
-	"math"
-	"math/rand"
 	"os"
 	"strconv"
-	"time"
 
+	"github.com/dwhitena/gophernet/gophernet"
 	"gonum.org/v1/gonum/floats"
 	"gonum.org/v1/gonum/mat"
 )
-
-// neuralNet contains all of the information
-// that defines a trained neural network.
-type neuralNet struct {
-	config  neuralNetConfig
-	wHidden *mat.Dense
-	bHidden *mat.Dense
-	wOut    *mat.Dense
-	bOut    *mat.Dense
-}
-
-// neuralNetConfig defines our neural network
-// architecture and learning parameters.
-type neuralNetConfig struct {
-	inputNeurons  int
-	outputNeurons int
-	hiddenNeurons int
-	numEpochs     int
-	learningRate  float64
-}
 
 func main() {
 
@@ -101,17 +78,17 @@ func main() {
 
 	// Define our network architecture and
 	// learning parameters.
-	config := neuralNetConfig{
-		inputNeurons:  4,
-		outputNeurons: 3,
-		hiddenNeurons: 3,
-		numEpochs:     5000,
-		learningRate:  0.3,
+	config := gophernet.GopherNetConfig{
+		InputNeurons:  4,
+		OutputNeurons: 3,
+		HiddenNeurons: 3,
+		NumEpochs:     5000,
+		LearningRate:  0.3,
 	}
 
 	// Train the neural network.
-	network := newNetwork(config)
-	if err := network.train(inputs, labels); err != nil {
+	network := gophernet.NewNetwork(config)
+	if err := network.Train(inputs, labels); err != nil {
 		log.Fatal(err)
 	}
 
@@ -178,7 +155,7 @@ func main() {
 	testLabels := mat.NewDense(len(rawCSVData), 3, labelsData)
 
 	// Make the predictions using the trained model.
-	predictions, err := network.predict(testInputs)
+	predictions, err := network.Predict(testInputs)
 	if err != nil {
 		log.Fatal(err)
 	}
@@ -209,202 +186,4 @@ func main() {
 
 	// Output the Accuracy value to standard out.
 	fmt.Printf("\nAccuracy = %0.2f\n\n", accuracy)
-}
-
-// NewNetwork initializes a new neural network.
-func newNetwork(config neuralNetConfig) *neuralNet {
-	return &neuralNet{config: config}
-}
-
-// train trains a neural network using backpropagation.
-func (nn *neuralNet) train(x, y *mat.Dense) error {
-
-	// Initialize biases/weights.
-	randSource := rand.NewSource(time.Now().UnixNano())
-	randGen := rand.New(randSource)
-
-	wHidden := mat.NewDense(nn.config.inputNeurons, nn.config.hiddenNeurons, nil)
-	bHidden := mat.NewDense(1, nn.config.hiddenNeurons, nil)
-	wOut := mat.NewDense(nn.config.hiddenNeurons, nn.config.outputNeurons, nil)
-	bOut := mat.NewDense(1, nn.config.outputNeurons, nil)
-
-	wHiddenRaw := wHidden.RawMatrix().Data
-	bHiddenRaw := bHidden.RawMatrix().Data
-	wOutRaw := wOut.RawMatrix().Data
-	bOutRaw := bOut.RawMatrix().Data
-
-	for _, param := range [][]float64{
-		wHiddenRaw,
-		bHiddenRaw,
-		wOutRaw,
-		bOutRaw,
-	} {
-		for i := range param {
-			param[i] = randGen.Float64()
-		}
-	}
-
-	// Define the output of the neural network.
-	output := new(mat.Dense)
-
-	// Use backpropagation to adjust the weights and biases.
-	if err := nn.backpropagate(x, y, wHidden, bHidden, wOut, bOut, output); err != nil {
-		return err
-	}
-
-	// Define our trained neural network.
-	nn.wHidden = wHidden
-	nn.bHidden = bHidden
-	nn.wOut = wOut
-	nn.bOut = bOut
-
-	return nil
-}
-
-// backpropagate completes the backpropagation method.
-func (nn *neuralNet) backpropagate(x, y, wHidden, bHidden, wOut, bOut, output *mat.Dense) error {
-
-	// Loop over the number of epochs utilizing
-	// backpropagation to train our model.
-	for i := 0; i < nn.config.numEpochs; i++ {
-
-		// Complete the feed forward process.
-		hiddenLayerInput := new(mat.Dense)
-		hiddenLayerInput.Mul(x, wHidden)
-		addBHidden := func(_, col int, v float64) float64 { return v + bHidden.At(0, col) }
-		hiddenLayerInput.Apply(addBHidden, hiddenLayerInput)
-
-		hiddenLayerActivations := new(mat.Dense)
-		applySigmoid := func(_, _ int, v float64) float64 { return sigmoid(v) }
-		hiddenLayerActivations.Apply(applySigmoid, hiddenLayerInput)
-
-		outputLayerInput := new(mat.Dense)
-		outputLayerInput.Mul(hiddenLayerActivations, wOut)
-		addBOut := func(_, col int, v float64) float64 { return v + bOut.At(0, col) }
-		outputLayerInput.Apply(addBOut, outputLayerInput)
-		output.Apply(applySigmoid, outputLayerInput)
-
-		// Complete the backpropagation.
-		networkError := new(mat.Dense)
-		networkError.Sub(y, output)
-
-		slopeOutputLayer := new(mat.Dense)
-		applySigmoidPrime := func(_, _ int, v float64) float64 { return sigmoidPrime(v) }
-		slopeOutputLayer.Apply(applySigmoidPrime, output)
-		slopeHiddenLayer := new(mat.Dense)
-		slopeHiddenLayer.Apply(applySigmoidPrime, hiddenLayerActivations)
-
-		dOutput := new(mat.Dense)
-		dOutput.MulElem(networkError, slopeOutputLayer)
-		errorAtHiddenLayer := new(mat.Dense)
-		errorAtHiddenLayer.Mul(dOutput, wOut.T())
-
-		dHiddenLayer := new(mat.Dense)
-		dHiddenLayer.MulElem(errorAtHiddenLayer, slopeHiddenLayer)
-
-		// Adjust the parameters.
-		wOutAdj := new(mat.Dense)
-		wOutAdj.Mul(hiddenLayerActivations.T(), dOutput)
-		wOutAdj.Scale(nn.config.learningRate, wOutAdj)
-		wOut.Add(wOut, wOutAdj)
-
-		bOutAdj, err := sumAlongAxis(0, dOutput)
-		if err != nil {
-			return err
-		}
-		bOutAdj.Scale(nn.config.learningRate, bOutAdj)
-		bOut.Add(bOut, bOutAdj)
-
-		wHiddenAdj := new(mat.Dense)
-		wHiddenAdj.Mul(x.T(), dHiddenLayer)
-		wHiddenAdj.Scale(nn.config.learningRate, wHiddenAdj)
-		wHidden.Add(wHidden, wHiddenAdj)
-
-		bHiddenAdj, err := sumAlongAxis(0, dHiddenLayer)
-		if err != nil {
-			return err
-		}
-		bHiddenAdj.Scale(nn.config.learningRate, bHiddenAdj)
-		bHidden.Add(bHidden, bHiddenAdj)
-	}
-
-	return nil
-}
-
-// predict makes a prediction based on a trained
-// neural network.
-func (nn *neuralNet) predict(x *mat.Dense) (*mat.Dense, error) {
-
-	// Check to make sure that our neuralNet value
-	// represents a trained model.
-	if nn.wHidden == nil || nn.wOut == nil {
-		return nil, errors.New("the supplied weights are empty")
-	}
-	if nn.bHidden == nil || nn.bOut == nil {
-		return nil, errors.New("the supplied biases are empty")
-	}
-
-	// Define the output of the neural network.
-	output := new(mat.Dense)
-
-	// Complete the feed forward process.
-	hiddenLayerInput := new(mat.Dense)
-	hiddenLayerInput.Mul(x, nn.wHidden)
-	addBHidden := func(_, col int, v float64) float64 { return v + nn.bHidden.At(0, col) }
-	hiddenLayerInput.Apply(addBHidden, hiddenLayerInput)
-
-	hiddenLayerActivations := new(mat.Dense)
-	applySigmoid := func(_, _ int, v float64) float64 { return sigmoid(v) }
-	hiddenLayerActivations.Apply(applySigmoid, hiddenLayerInput)
-
-	outputLayerInput := new(mat.Dense)
-	outputLayerInput.Mul(hiddenLayerActivations, nn.wOut)
-	addBOut := func(_, col int, v float64) float64 { return v + nn.bOut.At(0, col) }
-	outputLayerInput.Apply(addBOut, outputLayerInput)
-	output.Apply(applySigmoid, outputLayerInput)
-
-	return output, nil
-}
-
-// sigmoid implements the sigmoid function
-// for use in activation functions.
-func sigmoid(x float64) float64 {
-	return 1.0 / (1.0 + math.Exp(-x))
-}
-
-// sigmoidPrime implements the derivative
-// of the sigmoid function for backpropagation.
-func sigmoidPrime(x float64) float64 {
-	return x * (1.0 - x)
-}
-
-// sumAlongAxis sums a matrix along a
-// particular dimension, preserving the
-// other dimension.
-func sumAlongAxis(axis int, m *mat.Dense) (*mat.Dense, error) {
-
-	numRows, numCols := m.Dims()
-
-	var output *mat.Dense
-
-	switch axis {
-	case 0:
-		data := make([]float64, numCols)
-		for i := 0; i < numCols; i++ {
-			col := mat.Col(nil, i, m)
-			data[i] = floats.Sum(col)
-		}
-		output = mat.NewDense(1, numCols, data)
-	case 1:
-		data := make([]float64, numRows)
-		for i := 0; i < numRows; i++ {
-			row := mat.Row(nil, i, m)
-			data[i] = floats.Sum(row)
-		}
-		output = mat.NewDense(numRows, 1, data)
-	default:
-		return nil, errors.New("invalid axis, must be 0 or 1")
-	}
-
-	return output, nil
 }
